Move entity and config enums to Types.Enums
This is to help with organization mostly - and the fact that these enums could be used elsewhere and you may not want to reference the entire Slot entity for example to access that enum.
