+++ PLRPC/Types/Entities/Room.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using LBPUnion.PLRPC.Types.Enums;
 
 namespace LBPUnion.PLRPC.Types.Entities;
