+++ PLRPC/Types/Entities/UserStatus.cs
@@ -14,9 +14,3 @@ public class UserStatus
     [JsonPropertyName("currentRoom")]
     public Room? CurrentRoom { get; set; }
 }
-
-public enum StatusType
-{
-    Offline = 0,
-    Online = 1,
-}
