+++ PLRPC/Types/Enums/StatusType.cs
@@ -1,0 +1,7 @@
+﻿namespace LBPUnion.PLRPC.Types.Enums;
+
+public enum StatusType
+{
+    Offline = 0,
+    Online = 1,
+}
